requirements.txt
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.1
Pillow==9.5.0
psycopg2==2.9.6
django-cors-headers==3.14.0
django-filter==23.2
daphne==4.0.0
channels==4.0.0
celery==5.3.1
channels-redis==4.1.0